requirements.txt
werkzeug
requests
botocore
flask
boto3
python-dotenv
moto
pytest
gunicorn